feat: add debug header
